Re-arrange files, style board
