+++ components/Board/Board.module.scss
@@ -1,0 +1,5 @@
+.root {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: 32px;
+}
