+++ components/BoardSide/BoardSide.module.scss
